{"schema":"decodex/commit/1","summary":"Modularize macOS live stream facade","authority":"manual"}
